rename to engine; add basic resource manager
