+++ engine/events.go
@@ -1,4 +1,4 @@
-package lumen
+package engine
 
 import (
 	"fmt"
