+++ engine/sdl.go
@@ -1,4 +1,4 @@
-package lumen
+package engine
 
 import "github.com/veandco/go-sdl2/sdl"
